+++ src/Security/AclContext.php
@@ -12,7 +12,6 @@ final readonly class AclContext
         public array $groupIds = [],
         public bool $ticketTechnician = false,
         public bool $ticketRequester = false,
-        public bool $entityTechnician = false,
         public bool $itilItemAccess = false,
     ) {}
 }